fix(ink): address review — move test to tests/, for-agents docs, changeset, timer unref + EOL clear
- move progress-observer test to packages/ink/tests/ (src↔test parity gate)
- document createProgressObserver + SPINNER_FRAMES in for-agents/ink.mdx
- add changeset (core minor, ink minor)
- review fixes: unref() the spinner interval (no process hang on throw) + \x1b[K clear-to-EOL
